refactor: FunnelBasePage 이름 수정
