+++ frontend/src/pages/create/funnel/FunnelBasePage/FunnelBasePage.styles.ts
