+++ frontend/src/pages/create/funnelElements/TimeInputElement.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import DateTimeInput from '../../../components/@common/dateTimeInput/DateTimeInput';
 import type { FunnelElementProps } from '../../../types/funnel.type';
 import { checkIsPastDateTime } from '../../../utils/checkIsPastDateTime';
-import FunnelBasePage from '../funnel/funnelElementBase/FunnelElementBase';
+import FunnelBasePage from '../funnel/FunnelBasePage/FunnelBasePage';
 
 interface TimeInputElementProps extends FunnelElementProps {
   date: string;
